requirements.txt
fastapi
uvicorn
browser-use
playwright
python-dotenv
